+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/accountsecurity/accountrecoveryscan/AccountRecoveryScanSelectionViewModel.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.settings.recovery
+package com.babylon.wallet.android.presentation.settings.accountsecurity.accountrecoveryscan
 
 import androidx.lifecycle.viewModelScope
 import com.babylon.wallet.android.presentation.common.OneOffEvent
