+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/accountsecurity/accountrecoveryscan/chooseseed/ChooseSeedPhraseNav.kt
@@ -1,10 +1,7 @@
-@file:OptIn(ExperimentalAnimationApi::class)
-
-package com.babylon.wallet.android.presentation.settings.recovery.chooseseed
+package com.babylon.wallet.android.presentation.settings.accountsecurity.accountrecoveryscan.chooseseed
 
 import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.SavedStateHandle
 import androidx.navigation.NavController
@@ -15,7 +12,7 @@ import com.babylon.wallet.android.presentation.onboarding.restore.mnemonic.Mnemo
 import rdx.works.profile.data.model.factorsources.FactorSource
 
 private const val ARGS_MNEMONIC_TYPE = "mnemonic_type"
-private const val ROUTE = "choose_olympia_seed?$ARGS_MNEMONIC_TYPE={$ARGS_MNEMONIC_TYPE}"
+private const val ROUTE = "choose_seed_phrase?$ARGS_MNEMONIC_TYPE={$ARGS_MNEMONIC_TYPE}"
 
 internal class ChooseSeedPhraseArgs(val recoveryType: MnemonicType) {
     constructor(savedStateHandle: SavedStateHandle) : this(
@@ -28,7 +25,7 @@ internal class ChooseSeedPhraseArgs(val recoveryType: MnemonicType) {
 }
 
 fun NavController.chooseSeedPhrase(mnemonicType: MnemonicType) {
-    navigate(route = "choose_olympia_seed?$ARGS_MNEMONIC_TYPE=$mnemonicType")
+    navigate(route = "choose_seed_phrase?$ARGS_MNEMONIC_TYPE=$mnemonicType")
 }
 
 fun NavGraphBuilder.chooseSeedPhrase(
