+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/accountsecurity/accountrecoveryscan/chooseseed/ChooseSeedPhraseViewModel.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.settings.recovery.chooseseed
+package com.babylon.wallet.android.presentation.settings.accountsecurity.accountrecoveryscan.chooseseed
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
